Prepared for the assembly
