+++ src/scripts/service_goodbye_discord.bat
