+++ src/scripts/service_remove.bat
