+++ src/scripts/version.txt
